+++ src/clear.c
@@ -8,6 +8,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
+#include <stdio.h>
 
 #include "include/UART.h"
 
@@ -15,7 +16,7 @@
 
 #define TERM_CLEAR "\033[2J\033[H\033[3J" // The VT100 clear screen command
 
-uint8_t clear_entry(uint16_t argc, char **argv) {
+int32_t clear_entry(uint16_t argc, char **argv) {
     printf(TERM_CLEAR);
 
     return 0;
