+++ src/help.c
@@ -17,7 +17,7 @@
 
 
 
-uint8_t help_entry(uint16_t argc, char **argv) {
+int32_t help_entry(uint16_t argc, char **argv) {
     printf("Embedded Terminal version 0.1\r\n");
 
     printf("Use info embeddedTerminal for more infomation about this shell\r\n");
